refactor: extract resolveBuildConfig out of buildSchema
`buildSchema` was 467 lines, and about 400 of them were config resolution:
filling in prefix/suffix defaults, validating `limits`, `defaults`, `scope`,
`contextValues`, `softDelete`, `onWrite` and `exclude` against the real
schema, normalizing `complexity`, `transactions` and `eagerLoadRelations`,
and warning about feature-flag combinations — all of it before the code
knows which dialect it is talking to.

That half now lives in `src/util/build-config.ts` as
`resolveBuildConfig(config, schema): SchemaGeneratorOptions`. What is left
in `buildSchema` is what it is actually for: resolve the schema off the
drizzle instance, resolve the config, dispatch to the dialect generator,
install the error mapper, and assemble the `GraphQLSchema`.

- `src/index.ts`: 604 -> 187 lines; `buildSchema` 467 -> 71
- `src/util/build-config.ts`: +447

Pure code motion — the extracted block is unchanged apart from being
wrapped in a function that takes `config` and `schema` and returns the
options object it already built. Verified with `npm run typecheck` (both
projects), biome, the 70-file pglite/sqlite suite (1044 tests) and the
Docker pg/mysql suite (180 tests).
